Minor fix for small rectangular fields.
Changed track angle scoring to prefer directions
parallel to the longest field edge.

If we don't do this, we end up many angles with the same score
and the algorithm just took the first one. This resulted in
small rectangular fields with tracks in the center not parallel
with the field edge.